@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "engines": {
-    "node": "16.14.2"
+    "node": "22.14.0"
   },
   "dependencies": {
     "@headlessui/react": "^1.7.15",
@@ -19,17 +19,19 @@
     "@types/react-bootstrap": "^0.32.22",
     "@types/react-router-dom": "^5.3.3",
     "antd": "^4.7.3",
-    "bootstrap": "^4.6.1",
+    "bootstrap": "^5.3.3",
+    "buffer": "^6.0.3",
     "classnames": "^2.2.6",
     "date-fns": "^3.5.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
+    "events": "^3.3.0",
     "faker": "^5.4.0",
     "jquery": "^3.6.0",
     "jwt-decode": "^2.2.0",
     "normalize.css": "^8.0.1",
-    "postcss": "^8.4.26",
     "postcss-loader": "^7.0.0",
-    "react": "^18.0.0",
+    "process": "^0.11.10",
+    "react": "^18.3.1",
     "react-alert": "^7.0.0",
     "react-alert-template-basic": "^1.0.0",
     "react-anchor-link-smooth-scroll": "^1.0.12",
