@@ -64,7 +64,7 @@ export default function InviteSignupFlow({ orgName, personRole }: Props) {
   // @ts-ignore
   const currentStageIdx =
     signUpStageStateContext.stages?.indexOf(
-      signUpStageStateContext.currentStage
+      signUpStageStateContext.currentStage,
     ) || 0;
 
   return (
