@@ -80,7 +80,7 @@ export default function PersonSignupFlow() {
   // @ts-ignore
   const currentStageIdx =
     signUpStageStateContext.stages?.indexOf(
-      signUpStageStateContext.currentStage
+      signUpStageStateContext.currentStage,
     ) || 0;
 
   return (
