@@ -289,11 +289,11 @@ class LoginPage extends Component<Props, State> {
               </div>
               <div className="row pl-3 pb-3">
                 <span className="text-muted recaptcha-login-text pt-4 mt-4 pl-5 ml-5 w-75">
-                  This page is protected by reCAPTCHA, and subject to the Google
+                  This page is protected by reCAPTCHA, and subject to the Google{' '}
                   <a href="https://www.google.com/policies/privacy/">
                     Privacy Policy{' '}
                   </a>
-                  and
+                  and{' '}
                   <a href="https://www.google.com/policies/terms/">
                     Terms of service
                   </a>
@@ -388,7 +388,7 @@ class LoginPage extends Component<Props, State> {
                             )
                           }
                           onClick={this.handleSubmitTwoFactorCode}
-                          className={`mt-2 w-100 ld-ext-right ${buttonState}`}
+                          className={`mt-2 ld-ext-right ${buttonState}`}
                         >
                           Sign In
                           <div className="ld ld-ring ld-spin" />
@@ -412,7 +412,7 @@ class LoginPage extends Component<Props, State> {
                         }
                         onClick={this.onSubmitWithReCAPTCHA}
                         variant="primary"
-                        className={`px-5 w-100 ld-ext-right ${buttonState}`}
+                        className={`px-5 ld-ext-right ${buttonState}`}
                       >
                         Sign In
                         <div className="ld ld-ring ld-spin" />
