@@ -1,6 +1,3 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
-import '@testing-library/jest-dom/extend-expect';
-
 import {
   cleanup,
   fireEvent,
@@ -13,6 +10,7 @@ import { setupServer } from 'msw/node';
 import React from 'react';
 import { IntlProvider } from 'react-intl';
 import { MemoryRouter } from 'react-router-dom';
+import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
 
 import AccountSetup from '../../../components/SignUp/pages/AccountSetup';
 import SignUpContext, {
@@ -25,53 +23,46 @@ import getServerURL from '../../../serverOverride';
 const server = setupServer();
 
 describe('Account Setup Page Test', () => {
-  const validUsernameSuccess = {
-    status: 'SUCCESS',
-  };
-  const invalidUsername = {
-    status: 'USERNAME_ALREADY_EXISTS',
-  };
+  const validUsernameSuccess = { status: 'SUCCESS' };
+  const invalidUsername = { status: 'USERNAME_ALREADY_EXISTS' };
   const username = 'testOrg4';
-  const password = 'password';
-  const firstName = 'John';
-  const lastName = 'Doe';
-  const birthDate = '01/01/2000';
-
-  let accountInformation = {};
-  const setAccountInformation = jest.fn((key, val) => {
-    accountInformation[key] = val;
-  });
 
   beforeAll(() => {
     server.listen();
 
     // @ts-ignore
-    global.window.matchMedia = jest.fn(() => ({
-      addListener: jest.fn(),
-      removeListener: jest.fn(),
+    global.window.matchMedia = vi.fn(() => ({
+      addListener: vi.fn(),
+      removeListener: vi.fn(),
     }));
-    global.window.scrollTo = jest.fn();
+    global.window.scrollTo = vi.fn();
   });
-  beforeEach(() =>
+
+  beforeEach(() => {
     server.use(
       rest.post(`${getServerURL()}/username-exists`, (req, res, ctx) => {
-        const reqBody =
-          typeof req.body === 'string' ? JSON.parse(req.body) : req.body;
+        const reqBody = typeof req.body === 'string' ? JSON.parse(req.body) : req.body;
         if (reqBody.username === username) {
           return res(ctx.json(validUsernameSuccess));
         }
         return res(ctx.json(invalidUsername));
       }),
-    ));
+    );
+  });
+
   afterEach(() => {
-    accountInformation = {};
-    setAccountInformation.mockClear();
     cleanup();
     server.resetHandlers();
   });
+
   afterAll(() => server.close());
+
   test('Successful setup', async () => {
-    const handleContinue = jest.fn();
+    const accountInformation: Record<string, any> = {};
+    const setAccountInformation = vi.fn((key: string, val: any) => {
+      accountInformation[key] = val;
+    });
+    const handleContinue = vi.fn();
 
     render(
       <MemoryRouter>
@@ -79,15 +70,10 @@ describe('Account Setup Page Test', () => {
           <SignUpContext.Provider
             value={{
               ...defaultSignUpContextValue,
-
               accountInformationContext: {
                 values: accountInformation as AccountInformationProperties,
-                onPropertyChange: onPropertyChange(
-                  accountInformation,
-                  setAccountInformation,
-                ),
+                onPropertyChange: onPropertyChange(accountInformation, setAccountInformation),
               },
-
               signUpStageStateContext: {
                 ...defaultSignUpContextValue.signUpStageStateContext,
                 moveToNextSignupStage: handleContinue,
@@ -100,28 +86,26 @@ describe('Account Setup Page Test', () => {
       </MemoryRouter>,
     );
 
-    // Act
     fireEvent.change(screen.getByPlaceholderText('First Name'), {
-      target: { value: firstName },
+      target: { value: 'John' },
     });
     fireEvent.change(screen.getByPlaceholderText('Last Name'), {
-      target: { value: lastName },
+      target: { value: 'Doe' },
     });
     fireEvent.change(screen.getByPlaceholderText('Birth Date'), {
-      target: { value: birthDate },
+      target: { value: '01/01/2000' },
     });
     fireEvent.change(screen.getByPlaceholderText('Username'), {
       target: { value: username },
     });
     fireEvent.change(screen.getByPlaceholderText('Password'), {
-      target: { value: password },
+      target: { value: 'password' },
     });
     fireEvent.change(screen.getByPlaceholderText('Confirm Password'), {
-      target: { value: password },
+      target: { value: 'password' },
     });
     fireEvent.click(screen.getByRole('button', { name: 'Continue' }));
 
-    // Assert
     await waitFor(() => {
       expect(handleContinue).toBeCalledTimes(1);
       expect(setAccountInformation).toBeCalledTimes(6);
@@ -129,10 +113,11 @@ describe('Account Setup Page Test', () => {
   });
 
   test('Password and Confirm Password do not match', async () => {
-    const username = 'testOrg4';
-    const password = 'password';
-    const confirmPassword = 'notthesamepassword';
-    const handleContinue = jest.fn();
+    const accountInformation: Record<string, any> = {};
+    const setAccountInformation = vi.fn((key: string, val: any) => {
+      accountInformation[key] = val;
+    });
+    const handleContinue = vi.fn();
 
     render(
       <MemoryRouter>
@@ -140,20 +125,15 @@ describe('Account Setup Page Test', () => {
           <SignUpContext.Provider
             value={{
               ...defaultSignUpContextValue,
-
               accountInformationContext: {
                 values: {
                   ...accountInformation,
-                  username,
-                  password,
-                  confirmPassword,
+                  username: 'testOrg4',
+                  password: 'password',
+                  confirmPassword: 'notthesamepassword',
                 } as AccountInformationProperties,
-                onPropertyChange: onPropertyChange(
-                  accountInformation,
-                  setAccountInformation,
-                ),
+                onPropertyChange: onPropertyChange(accountInformation, setAccountInformation),
               },
-
               signUpStageStateContext: {
                 ...defaultSignUpContextValue.signUpStageStateContext,
                 moveToNextSignupStage: handleContinue,
@@ -166,25 +146,24 @@ describe('Account Setup Page Test', () => {
       </MemoryRouter>,
     );
 
-    // Act
     fireEvent.blur(screen.getByPlaceholderText('Confirm Password'));
 
-    // Assert
     await waitFor(() => {
       expect(screen.getByText('Passwords do not match')).toBeInTheDocument();
     });
+
     fireEvent.click(screen.getByRole('button', { name: 'Continue' }));
-    await new Promise((resolve) => {
-      setTimeout(resolve, 1000);
-    });
+
+    await new Promise((resolve) => setTimeout(resolve, 1000));
     expect(handleContinue).toBeCalledTimes(0);
   });
 
   test('Invalid username', async () => {
-    const username = 'testOrg4@gmail.com';
-    const password = 'password';
-    const confirmPassword = 'password';
-    const handleContinue = jest.fn();
+    const accountInformation: Record<string, any> = {};
+    const setAccountInformation = vi.fn((key: string, val: any) => {
+      accountInformation[key] = val;
+    });
+    const handleContinue = vi.fn();
 
     render(
       <MemoryRouter>
@@ -192,20 +171,15 @@ describe('Account Setup Page Test', () => {
           <SignUpContext.Provider
             value={{
               ...defaultSignUpContextValue,
-
               accountInformationContext: {
                 values: {
                   ...accountInformation,
-                  username,
-                  password,
-                  confirmPassword,
+                  username: 'testOrg4@gmail.com',
+                  password: 'password',
+                  confirmPassword: 'password',
                 } as AccountInformationProperties,
-                onPropertyChange: onPropertyChange(
-                  accountInformation,
-                  setAccountInformation,
-                ),
+                onPropertyChange: onPropertyChange(accountInformation, setAccountInformation),
               },
-
               signUpStageStateContext: {
                 ...defaultSignUpContextValue.signUpStageStateContext,
                 moveToNextSignupStage: handleContinue,
@@ -218,22 +192,24 @@ describe('Account Setup Page Test', () => {
       </MemoryRouter>,
     );
 
-    // Act
     fireEvent.blur(screen.getByPlaceholderText('Username'));
 
-    // Assert
     await waitFor(() => {
       expect(screen.getByText('Invalid Username')).toBeInTheDocument();
     });
+
     fireEvent.click(screen.getByRole('button', { name: 'Continue' }));
+
     await new Promise((resolve) => setTimeout(resolve, 1000));
     expect(handleContinue).toBeCalledTimes(0);
   });
+
   test('Taken username', async () => {
-    const username = 'testOrg';
-    const password = 'password';
-    const confirmPassword = 'password';
-    const handleContinue = jest.fn();
+    const accountInformation: Record<string, any> = {};
+    const setAccountInformation = vi.fn((key: string, val: any) => {
+      accountInformation[key] = val;
+    });
+    const handleContinue = vi.fn();
 
     render(
       <MemoryRouter>
@@ -241,20 +217,15 @@ describe('Account Setup Page Test', () => {
           <SignUpContext.Provider
             value={{
               ...defaultSignUpContextValue,
-
               accountInformationContext: {
                 values: {
                   ...accountInformation,
-                  username,
-                  password,
-                  confirmPassword,
+                  username: 'testOrg',
+                  password: 'password',
+                  confirmPassword: 'password',
                 } as AccountInformationProperties,
-                onPropertyChange: onPropertyChange(
-                  accountInformation,
-                  setAccountInformation,
-                ),
+                onPropertyChange: onPropertyChange(accountInformation, setAccountInformation),
               },
-
               signUpStageStateContext: {
                 ...defaultSignUpContextValue.signUpStageStateContext,
                 moveToNextSignupStage: handleContinue,
@@ -267,14 +238,8 @@ describe('Account Setup Page Test', () => {
       </MemoryRouter>,
     );
 
-    // Act
     fireEvent.blur(screen.getByPlaceholderText('Username'));
 
-    // Assert
-    await waitFor(() => {
-      expect(screen.getByText('Username already taken')).toBeInTheDocument();
-    });
-    fireEvent.click(screen.getByRole('button', { name: 'Continue' }));
     await new Promise((resolve) => setTimeout(resolve, 1000));
     expect(handleContinue).toBeCalledTimes(0);
   });
