@@ -6,15 +6,22 @@ module.exports = withMT({
     preflight: false,
   },
   prefix: 'tw-',
-  content: ['./src/**/*.{js,jsx,ts,tsx}'],
+  // 1) Add './index.html' and `.{scss}` if you use classes in .scss
+  content: [
+    './index.html',
+    './src/**/*.{js,jsx,ts,tsx,scss}',
+  ],
   theme: {
     extend: {
-      // ...
       fontFamily: {
         body: ['Inter'],
+        raleway: ['Raleway', 'sans-serif'],
       },
       colors: {
         twprimary: '#445FEB',
+        'primary-theme': '#445feb',
+        'secondary-text-color': '#b7b7c0',
+        'secondary-theme': '#413f57',
         footerblack: '#1D1D1D',
       },
     },
